requirements.txt
streamlit
pytesseract
pydantic
opencv-python
pillow
pandas
matplotlib